Add Helpers class
